Restructured files and README.md changes.
