+++ src/draw-and-plot-batch.js
@@ -4,7 +4,7 @@ var Lfullscreen = require('leaflet-fullscreen')
 var MiniMap = require('leaflet-minimap');
 //var markercluster = require('leaflet-markercluster');
 
-var utils = require('./utils.js');
+var drawControlimport = require('../utils.js');
 
 L.Icon.Default.imagePath = 'node_modules/leaflet/dist/images/';
 
@@ -94,7 +94,8 @@ map.on("draw:created", function (e) {
 });
 //////////////////////////END DRAW CONTROL ON THE MAP
 
-//fullscreen control
+
+//////////////////////////start fullscreen CONTROL ON THE MAP
 map.addControl(new L.Control.Fullscreen());
 map.on('fullscreenchange', function () {
     if (map.isFullscreen()) {
@@ -106,16 +107,16 @@ map.on('fullscreenchange', function () {
         fscrinfo.remove();
     }
 });
+//////////////////////////end fullscreen CONTROL ON THE MAP
 
-
-//Minimap control
-var osmUrl = "http://server.arcgisonline.com/ArcGIS/rest/services/NatGeo_World_Map/MapServer/tile/{z}/{y}/{x}"; //"http://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png";
-var osmAttrib = "Map data &copy; OpenStreetMap contributors";
-//Minimap Plugin magic goes here! Note that you cannot use the same layer object again, as that will confuse the two map controls
-var osm2 = new L.TileLayer(osmUrl, { minZoom: 0, maxZoom: 11 });
-var minimap_options={ toggleDisplay: true, width: 200, height: 200, collapsedHeight:19,collapsedWidth:19,zoomOffset:map.getZoom()-1 }
-var miniMap = new L.Control.MiniMap(osm2,minimap_options).addTo(map);
-// new MiniMap(layer, options).addTo(map);
+// //////////////////////////START Minimap CONTROL ON THE MAP
+// var osmUrl = "http://server.arcgisonline.com/ArcGIS/rest/services/NatGeo_World_Map/MapServer/tile/{z}/{y}/{x}"; //"http://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png";
+// var osmAttrib = "Map data &copy; OpenStreetMap contributors";
+// //Minimap Plugin magic goes here! Note that you cannot use the same layer object again, as that will confuse the two map controls
+// var osm2 = new L.TileLayer(osmUrl, { minZoom: 0, maxZoom: 11 });
+// var minimap_options={ toggleDisplay: true, width: 200, height: 200, collapsedHeight:19,collapsedWidth:19,zoomOffset:map.getZoom()-1 }
+// var miniMap = new L.Control.MiniMap(osm2,minimap_options).addTo(map);
+// //////////////////////////END Minimap CONTROL ON THE MAP
 
 /*Performance check by adding large  number of circle markers*/
 // control that shows state info on hover
@@ -133,38 +134,25 @@ info.update = function (props,state="") {
 
 info.addTo(map);
 
-const allCircles = [];
-var circlecounter = 0;
-const startTime = Date.now();
-const circlesUpdater = setInterval(() => {
-    allCircles.forEach(circleMeta => {
-    });
-
-    updateProps = {
-        name: allCircles.length,
-        timestamp: Date.now() - startTime,
-    };
-    info.update(updateProps);
-
-    if (circlecounter > 2000) {
-        clearInterval(circlesUpdater);
+const starttime = Date.now();
+var preplotpoints = [];
+for(i = 0; i < 250000;i++){
+    var point = drawControlimport.getRandomLatLng(map)
+    if(preplotpoints.indexOf(point) === -1) preplotpoints.push(point);
+    var marker = new L.circle(preplotpoints[i]);
+    map.addLayer(marker);
+    if(i % 1000 === 0){
+        console.log(i);
     }
-    const newCircle = L.circle(utils.getRandomLatLng(map), 50, {
-        color: 'red',
-        fillColor: '#f03',
-        fillOpacity: 0.5
-    }).addTo(map);
+}
+updateProps = {
+    name: preplotpoints.length,
+    timestamp: Date.now() - starttime,
+};
+info.update(updateProps);
+console.log(" plotting "+ preplotpoints.length.toString() +"stored points takes" + (Date.now() - starttime).toString() +"ms ");
+console.log(" plotting "+ preplotpoints.length.toString() +"stored points takes" + performance.memory.usedJSHeapSize.toString + "ms ");
 
-    circlecounter++;
-
-    allCircles.push({
-        circle: newCircle,
-        theta: 1// (Math.random() * 2) - 1,
-    });
-    if(allCircles.length % 100 == 0){
-        console.log("Plotting " + allCircles.length.toString() + " takes "+ (Date.now() - startTime).toString() + "ms" );
-    }
-},5)
 
 module.exports = {
     map
